chore(ui): adjust version list styles
